% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masses_1.R
\docType{data}
\name{masses_1}
\alias{masses_1}
\title{Masses from example 1}
\format{
A data frame with 20 rows and 6 columns:
\describe{
\item{sample}{sample number}
\item{extract}{mass of water sample used for derivatization}
\item{spike_IHg}{mass of inorganic quantification spike, in our case enriched in 198 isotope}
\item{spike_MMHg}{mass of monomethylmercury quantification spike, in our case enriched in 202 isotope}
...
}
}
\source{
{created in-house to serve as an example}
}
\usage{
masses_1
}
\description{
This table provided the all the masses used during sample preparation.
The masses are provided in grams and are used to correct the concentrations in case there are deviations when measuring or pipeting.
In the provided example you can find: sample number, ,
}
\examples{
masses_1
}
\keyword{datasets}
