% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_km.R
\name{mean_km}
\alias{mean_km}
\title{Calculations of Hg the methylation rates}
\usage{
mean_km(
  x,
  y,
  z,
  samples,
  injections = 3,
  type = c("liquid", "solid"),
  incub_IHg,
  incub_MMHg,
  quant_MMHg,
  quant_IHg,
  nat,
  spike_concentration_MMHg,
  spike_concentration_IHg
)
}
\arguments{
\item{x}{data frame with your integrated peaks}

\item{y}{data frame with your masses}

\item{z}{data frame with your abundances}

\item{samples}{Numeric vector. Number of samples}

\item{injections}{Numeric vector. Number of injections per sample Default = 3}

\item{type}{Type of sample: liquid or solid}

\item{incub_IHg}{IHg isotope used for the incubation}

\item{incub_MMHg}{MMHg isotope used for the incubation}

\item{quant_MMHg}{MMHg isotope used for the quantification}

\item{quant_IHg}{IHg isotope used for the quantification}

\item{nat}{isotope used as natural for the incubation}

\item{spike_concentration_MMHg}{Concentration of the organic Hg quantification spike. Units: ng/g}

\item{spike_concentration_IHg}{Concentration of the inorganic Hg quantification spike. Units: ng/g}
}
\value{
Table with the sample mean Hg methylation rates and the sd.
}
\description{
Calculations of Hg the methylation rates
}
\examples{
mean_km(
  peaks_3,
  masses_3,
  abundances_3,
  samples = 96,
  injections =3,
  type = 'solid',
  incub_IHg=199,
  incub_MMHg=201,
  quant_MMHg = 202,
  quant_IHg= 198,
  nat=200,
  spike_concentration_MMHg=0.5,
  spike_concentration_IHg=5)
 mean_km(
  peaks_1,
  masses_1,
  abundances_1,
  samples = 24,
  injections = 3,
  type = 'liquid',
  incub_IHg=199,
  incub_MMHg=201,
  quant_MMHg = 202,
  quant_IHg= 198,
  nat=200,
  spike_concentration_MMHg=0.5,
  spike_concentration_IHg=5)
  mean_km(
  peaks_2,
  masses_2,
  abundances_2,
  samples = 26,
  injections = 3,
  type = 'solid',
  incub_IHg=204,
  incub_MMHg=200,
  quant_MMHg = 201,
  quant_IHg= 199,
  nat=202,
  spike_concentration_MMHg=0.23,
  spike_concentration_IHg=27.27)
}
