% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMHg_byinjection_2.R
\name{MMHg}
\alias{MMHg}
\title{Calculates concentration of the inorganic (MMHg) isotopes in sediments}
\usage{
MMHg(
  x,
  y,
  z,
  samples,
  injections = 3,
  type = c("liquid", "solid"),
  quant_MMHg,
  spike_concentration_MMHg
)
}
\arguments{
\item{x}{data frame with your integrated peaks}

\item{y}{data frame with your masses}

\item{z}{data frame with your abundances}

\item{samples}{Numeric vector. Number of samples}

\item{injections}{Numeric vector. Number of injections per sample Default = 3}

\item{type}{Type of sample: liquid or solid}

\item{quant_MMHg}{MMHg isotope used for the quantification}

\item{spike_concentration_MMHg}{Concentration of the organic Hg quantification spike. Units: ng/g}
}
\value{
Table with the mean sample concentrations and the standard deviations for each isotope.
}
\description{
The mean concentration of MMHg and their standard deviations of the different isotopes are returned in a table. Mean and sd are calculated with the injection values.
}
\examples{
MMHg(
x=peaks_3,
  y=masses_3,
  z=abundances_3,
  samples = 96,
  type = 'solid',
quant_MMHg = 202,
spike_concentration_MMHg = 5)
MMHg(peaks_1, masses_1, abundances_1, 24, 3, type = 'liquid' , 198, 5)
MMHg(peaks_2, masses_2, abundances_2, 26, 3, type = 'solid' ,199, 27.27)
}
