% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_2.R
\docType{data}
\name{peaks_2}
\alias{peaks_2}
\title{Peaks table used for example 2}
\format{
A data frame with 20 rows and 12 columns
\describe{
\item{sample}{sample number}
\item{injection}{injection replicate, in our case we did triplicates; values 1-3}
\item{ MMHg199, MMHg200, MMHg201, MMHg202, MMHg204,
IHg199, IHg200, IHg201, IHg202, IHg204}{value of the integraded area from the GC-ICP-MS peaks for each isotope and form}
...
}
}
\source{
{created in-house to serve as an example}
}
\usage{
peaks_2
}
\description{
This table provided the all the peaks integrated from the results of the GC-ICP-MS,
all MMHg and IHg are included in the same file to facilitate the prior construction of the dataset.
}
\examples{
peaks_2
}
\keyword{datasets}
