% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_3.R
\docType{data}
\name{peaks_3}
\alias{peaks_3}
\title{Peaks table used for example 3}
\format{
A data frame with 20 rows and 12 columns
\describe{
\item{sample}{sample number}
\item{injection}{injection replicate, in our case we did triplicates; values 1-3}
\item{MMHg196, IHg196, MMHg198,IHg198,  MMHg199,IHg199,  MMHg200, IHg200, MMHg201, IHg201,  MMHg202, IHg202, MMHg204, IHg204}{value of the integraded area from the GC-ICP-MS peaks for each isotope and form}
...
}
}
\source{
{created in-house to serve as an example}
}
\usage{
peaks_3
}
\description{
This table provided the all the peaks integrated from the results of the GC-ICP-MS,
all MMHg and IHg are included in the same file to facilitate the prior construction of the dataset.
}
\examples{
peaks_3
}
\keyword{datasets}
