% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsd_peaks.R
\name{rsd_peaks}
\alias{rsd_peaks}
\title{Check the quality of your integrations using the relative standard deviation of your peak's ratios}
\usage{
rsd_peaks(x, form = c("IHg", "MMHg", "all"), cutoff = 10)
}
\arguments{
\item{x}{data frame with your integrated peaks}

\item{form}{IHg, MMHg, all}

\item{cutoff}{Numeric value. cutoff used for the plot. Default = 10\%}
}
\value{
table with the area peaks RSD
}
\description{
A short description...
}
\examples{
rsd_peaks(peaks_1, form = 'MMHg')
rsd_peaks(peaks_2, form = 'all', cutoff = 2)
rsd_peaks(peaks_3, form = 'IHg', 5)
}
