% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsd_ratios.R
\name{rsd_ratios}
\alias{rsd_ratios}
\title{Check the quality of your integrations using the relative standard deviation of your peak's ratios}
\usage{
rsd_ratios(
  x,
  form = c("IHg", "MMHg", "all"),
  samples,
  cutoff = 10,
  incub_IHg,
  incub_MMHg,
  quant_IHg,
  quant_MMHg,
  nat
)
}
\arguments{
\item{x}{data frame with your integrated peaks}

\item{form}{IHg, MMHg, all}

\item{samples}{Numeric vector. Number of samples}

\item{cutoff}{Numeric value. cutoff used for the plot. Default = 10\%}

\item{incub_IHg}{IHg isotope used for the incubation}

\item{incub_MMHg}{MMHg isotope used for the incubation}

\item{quant_IHg}{IHg isotope used for the quantification}

\item{quant_MMHg}{MMHg isotope used for the quantification}

\item{nat}{isotope used as natural for the incubation}
}
\value{
a table and one plot per sample of the ones that are above the cutoff.
}
\description{
A short description...
}
\examples{
rsd_ratios(peaks_1,
form = 'MMHg',
sample = 24,
incub_IHg = 199,
incub_MMHg = 201,
quant_IHg = 198,
quant_MMHg = 202,
nat = 200)
rsd_ratios(peaks_2,
form = 'all',
sample = 26,
incub_IHg = 200,
incub_MMHg = 204,
quant_IHg = 199,
quant_MMHg = 201,
nat = 202)
rsd_ratios(peaks_3, form = 'MMHg', 96, 5, 199, 201, 198, 202, 200)
}
