% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMHg_byinjection_2.R
\name{MMHg}
\alias{MMHg}
\title{Calculates concentration of the inorganic (MMHg) isotopes in sediments}
\usage{
MMHg(
  x,
  y,
  z,
  samples,
  injections = 3,
  type = c("liquid", "solid"),
  incub_IHg,
  incub_MMHg,
  quant_IHg,
  quant_MMHg,
  nat,
  spike_concentration_MMHg
)
}
\arguments{
\item{x}{data frame with your integrated peaks}

\item{y}{data frame with your masses}

\item{z}{data frame with your abundances}

\item{samples}{Numeric vector. Number of samples}

\item{injections}{Numeric vector. Number of injections per sample Default = 3}

\item{type}{Type of sample: liquid or solid}

\item{incub_IHg}{IHg isotope used for the incubation}

\item{incub_MMHg}{MMHg isotope used for the incubation}

\item{quant_IHg}{IHg isotope used for the quantification}

\item{quant_MMHg}{MMHg isotope used for the quantification}

\item{nat}{isotope used as natural for the incubation}

\item{spike_concentration_MMHg}{Concentration of the organic Hg quantification spike. Units: ng/g}
}
\value{
Table with the mean sample concentrations and the standard deviations for each isotope.
}
\description{
The mean concentration of MMHg and their standard deviations of the different isotopes are returned in a table. Mean and sd are calculated with the injection values.
}
\examples{
MMHg(
x=peaks_3,
  y=masses_3,
  z=abundances_3,
  samples = 96,
  type = 'solid',
incub_IHg = 199,
incub_MMHg=201,
quant_MMHg=202,
quant_IHg=198,
nat=200,
spike_concentration_MMHg = 0.5)
MMHg(peaks_1, masses_1, abundances_1, 24, 3, type = 'liquid' ,199,201,198,202,200, 0.22)
MMHg(peaks_2, masses_2, abundances_2, 26, 3, type = 'solid' ,204,200,199,201,202,0.23)
}
